Add finish filter refactor for java

+++ project/jimmer-sql/src/test/java/org/babyfish/jimmer/sql/filter/SaveWithTriggerTest.java
@@ -2,20 +2,25 @@ package org.babyfish.jimmer.sql.filter;
 
 import org.babyfish.jimmer.sql.JSqlClient;
 import org.babyfish.jimmer.sql.common.AbstractMutationTest;
+import org.babyfish.jimmer.sql.event.TriggerType;
 import org.babyfish.jimmer.sql.filter.common.FileFilter;
 import org.babyfish.jimmer.sql.meta.UserIdGenerator;
 import org.babyfish.jimmer.sql.model.filter.File;
 import org.babyfish.jimmer.sql.model.filter.FileDraft;
 import org.babyfish.jimmer.sql.model.filter.User;
+import org.babyfish.jimmer.sql.trigger.AbstractTriggerTest;
 import org.junit.jupiter.api.Test;
 
-public class SaveTest extends AbstractMutationTest {
+import java.util.List;
+
+public class SaveWithTriggerTest extends AbstractTriggerTest {
 
     private final JSqlClient sqlClient =
             getSqlClient(it -> {
                 UserIdGenerator<?> idGenerator = this::autoId;
                 it.setIdGenerator(idGenerator);
                 it.addFilters(new FileFilter());
+                it.setTriggerType(TriggerType.TRANSACTION_ONLY);
             });
 
     @Test
@@ -49,9 +54,14 @@ public class SaveTest extends AbstractMutationTest {
                         });
                         ctx.statement(it -> {
                             it.sql(
-                                    "update FILE set PARENT_ID = ? where ID = ?"
+                                    "select tb_1_.ID, tb_1_.NAME, tb_1_.PARENT_ID " +
+                                            "from FILE tb_1_ where tb_1_.ID = ? and exists(" +
+                                            "--->select 1 " +
+                                            "--->from FILE_USER_MAPPING tb_2_ " +
+                                            "--->where tb_2_.FILE_ID = tb_1_.ID and tb_2_.USER_ID = ?" +
+                                            ")"
                             );
-                            it.variables(8L, 9L);
+                            it.variables(9L, 2L);
                         });
                         ctx.statement(it -> {
                             it.sql(
@@ -85,20 +95,28 @@ public class SaveTest extends AbstractMutationTest {
                         });
                         ctx.statement(it -> {
                             it.sql(
-                                    "delete from FILE_USER_MAPPING where FILE_ID in (?, ?)"
+                                    "select FILE_ID, USER_ID from FILE_USER_MAPPING where FILE_ID in (?, ?)"
                             );
                             it.variables(11L, 12L);
                         });
                         ctx.statement(it -> {
                             it.sql(
-                                    "select ID from FILE where PARENT_ID in (?, ?)"
+                                    "delete from FILE_USER_MAPPING where (FILE_ID, USER_ID) in ((?, ?), (?, ?), (?, ?), (?, ?))"
+                            );
+                            it.variables(11L, 2L, 11L, 4L, 12L, 2L, 12L, 3L);
+                        });
+                        ctx.statement(it -> {
+                            it.sql(
+                                    "select tb_1_.ID, tb_1_.NAME, tb_1_.PARENT_ID from FILE tb_1_ where tb_1_.PARENT_ID in (?, ?)"
                             );
                             it.variables(11L, 12L);
                         });
                         ctx.statement(it -> {
-                            it.sql(
-                                    "delete from FILE where ID in (?, ?)"
-                            );
+                            it.sql("select tb_1_.ID, tb_1_.NAME, tb_1_.PARENT_ID from FILE tb_1_ where tb_1_.ID in (?, ?)");
+                            it.variables(11L, 12L);
+                        });
+                        ctx.statement(it -> {
+                            it.sql("delete from FILE where ID in (?, ?)");
                             it.variables(11L, 12L);
                         });
                         ctx.entity(it -> {
@@ -116,6 +134,33 @@ public class SaveTest extends AbstractMutationTest {
                     }
             );
         });
+        assertEvents(
+                "Event{" +
+                        "--->oldEntity=null, " +
+                        "--->newEntity={" +
+                        "--->--->\"id\":10000," +
+                        "--->--->\"name\":\"new_file\"," +
+                        "--->--->\"parent\":{\"id\":8}" +
+                        "--->}, " +
+                        "--->reason=null" +
+                        "}",
+                "AssociationEvent{prop=org.babyfish.jimmer.sql.model.filter.File.parent, sourceId=10000, detachedTargetId=null, attachedTargetId=8, reason=null}",
+                "AssociationEvent{prop=org.babyfish.jimmer.sql.model.filter.File.childFiles, sourceId=8, detachedTargetId=null, attachedTargetId=10000, reason=null}",
+                "AssociationEvent{prop=org.babyfish.jimmer.sql.model.filter.File.users, sourceId=11, detachedTargetId=2, attachedTargetId=null, reason=null}",
+                "AssociationEvent{prop=org.babyfish.jimmer.sql.model.filter.User.files, sourceId=2, detachedTargetId=11, attachedTargetId=null, reason=null}",
+                "AssociationEvent{prop=org.babyfish.jimmer.sql.model.filter.File.users, sourceId=11, detachedTargetId=4, attachedTargetId=null, reason=null}",
+                "AssociationEvent{prop=org.babyfish.jimmer.sql.model.filter.User.files, sourceId=4, detachedTargetId=11, attachedTargetId=null, reason=null}",
+                "AssociationEvent{prop=org.babyfish.jimmer.sql.model.filter.File.users, sourceId=12, detachedTargetId=2, attachedTargetId=null, reason=null}",
+                "AssociationEvent{prop=org.babyfish.jimmer.sql.model.filter.User.files, sourceId=2, detachedTargetId=12, attachedTargetId=null, reason=null}",
+                "AssociationEvent{prop=org.babyfish.jimmer.sql.model.filter.File.users, sourceId=12, detachedTargetId=3, attachedTargetId=null, reason=null}",
+                "AssociationEvent{prop=org.babyfish.jimmer.sql.model.filter.User.files, sourceId=3, detachedTargetId=12, attachedTargetId=null, reason=null}",
+                "Event{oldEntity={\"id\":11,\"name\":\"purge\",\"parent\":{\"id\":8}}, newEntity=null, reason=null}",
+                "AssociationEvent{prop=org.babyfish.jimmer.sql.model.filter.File.parent, sourceId=11, detachedTargetId=8, attachedTargetId=null, reason=null}",
+                "AssociationEvent{prop=org.babyfish.jimmer.sql.model.filter.File.childFiles, sourceId=8, detachedTargetId=11, attachedTargetId=null, reason=null}",
+                "Event{oldEntity={\"id\":12,\"name\":\"ssh\",\"parent\":{\"id\":8}}, newEntity=null, reason=null}",
+                "AssociationEvent{prop=org.babyfish.jimmer.sql.model.filter.File.parent, sourceId=12, detachedTargetId=8, attachedTargetId=null, reason=null}",
+                "AssociationEvent{prop=org.babyfish.jimmer.sql.model.filter.File.childFiles, sourceId=8, detachedTargetId=12, attachedTargetId=null, reason=null}"
+        );
     }
 
     @Test
@@ -144,7 +189,7 @@ public class SaveTest extends AbstractMutationTest {
                         });
                         ctx.statement(it -> {
                             it.sql(
-                                    "select tb_1_.ID, tb_1_.NAME " +
+                                    "select tb_1_.ID, tb_1_.NAME, tb_1_.DELETED_TIME " +
                                             "from file_user tb_1_ " +
                                             "where tb_1_.NAME = ? and tb_1_.DELETED_TIME is null"
                             );
@@ -192,5 +237,17 @@ public class SaveTest extends AbstractMutationTest {
                     }
             );
         });
+        
+        assertEvents(
+                "Event{oldEntity=null, newEntity={\"id\":10000,\"name\":\"Andrew\"}, reason=null}",
+                "AssociationEvent{prop=org.babyfish.jimmer.sql.model.filter.File.users, sourceId=20, detachedTargetId=2, attachedTargetId=null, reason=null}",
+                "AssociationEvent{prop=org.babyfish.jimmer.sql.model.filter.User.files, sourceId=2, detachedTargetId=20, attachedTargetId=null, reason=null}",
+                "AssociationEvent{prop=org.babyfish.jimmer.sql.model.filter.File.users, sourceId=20, detachedTargetId=4, attachedTargetId=null, reason=null}",
+                "AssociationEvent{prop=org.babyfish.jimmer.sql.model.filter.User.files, sourceId=4, detachedTargetId=20, attachedTargetId=null, reason=null}",
+                "AssociationEvent{prop=org.babyfish.jimmer.sql.model.filter.File.users, sourceId=20, detachedTargetId=null, attachedTargetId=3, reason=null}",
+                "AssociationEvent{prop=org.babyfish.jimmer.sql.model.filter.User.files, sourceId=3, detachedTargetId=null, attachedTargetId=20, reason=null}",
+                "AssociationEvent{prop=org.babyfish.jimmer.sql.model.filter.File.users, sourceId=20, detachedTargetId=null, attachedTargetId=10000, reason=null}",
+                "AssociationEvent{prop=org.babyfish.jimmer.sql.model.filter.User.files, sourceId=10000, detachedTargetId=null, attachedTargetId=20, reason=null}"
+        );
     }
 }
